Significantly reduce cost of large presets like `Name Suggestion Index`
This reduces the startup CPU of TaggingPresets#initialize by ~29% and the memory
by ~55%. This was done by adding a batch `addActiveLayerChangeListeners` method
in MainLayerManager.

This additionally reduces the cost of
`TaggingPresetSelect$PresetClassification.simplifyString` (used by the preset
search dialog) by ~28% CPU cycles and ~58% memory allocations. This was done by
having a single static Pattern which was reused -- most of the remaining
potentially reducible cost for simplifyString is from Matcher initialization.
This would require thread local variables, so may not be worthwhile for now.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18691 0c6e7542-c601-0410-84e7-c038aed88b3b