Fix #22990: Synchronize ivy.xml with the actual resolved artifact
jakarta.json 2.1.1 was compiled for Java 8, except the package-info files. As
such, even though the specification page indicates that it is Java 11 only, it
does, in fact, work with Java 8.

parsson 1.1.1 requires jakarta.json 2.1.1, and overrode the version pinned in
ivy, which led to compile-time issues for the Debian Project package.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18750 0c6e7542-c601-0410-84e7-c038aed88b3b