Reduce allocations during startup from Shape
Shape#hashCode made between 8 and 24 MB of allocations at startup. This was
reduced by moving the math out of Arrays#hash and the default List#hash methods.
This additionally reduced the allocations from the Shape#equals method.

As an implementation note, the Shape#hashCode method returns exactly the same
hashcode as it did prior to this change. This may change in the future.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18777 0c6e7542-c601-0410-84e7-c038aed88b3b