See #23218: Use newer error_prone versions when compiling on Java 11+
error_prone 2.11 dropped support for compiling with Java 8, although it still
supports compiling for Java 8. The "major" new check for us is `NotJavadoc` since
we used `/**` in quite a few places which were not javadoc.

Other "new" checks that are of interest:
* `AlreadyChecked`: `if (foo) { doFoo(); } else if (!foo) { doBar(); }`
* `UnnecessaryStringBuilder`: Avoid `StringBuilder` (Java converts `+` to
  `StringBuilder` behind-the-scenes, but may also do something else if it
  performs better)
* `NonApiType`: Avoid specific interface types in function definitions
* `NamedLikeContextualKeyword`: Avoid using restricted names for classes and
  methods
* `UnusedMethod`: Unused private methods should be removed

This fixes most of the new error_prone issues and some SonarLint issues.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18871 0c6e7542-c601-0410-84e7-c038aed88b3b