Fix #16567: Upgrade to JUnit 5
JOSMTestRules and JOSMTestFixture can reset the default JOSM profile, which can
be unexpected for new contributors. This updates all tests to use JUnit 5 ''and''
the new JUnit 5 annotations.

This also renames MapCSSStyleSourceFilterTest to MapCSSStyleSourceFilterPerformanceTest
to match the naming convention for performance tests and fixes some lint issues.

This was tested by running all tests individually and together.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18893 0c6e7542-c601-0410-84e7-c038aed88b3b